Replace VerifyEmailPayload with UserLogInPayload
