+++ packages/server/graphql/public/permissions.ts
@@ -37,6 +37,7 @@ const permissionMap: PermissionMap<Resolvers> = {
       not(isEnvVarTrue('AUTH_INTERNAL_DISABLED')),
       rateLimit({perMinute: 50, perHour: 500})
     ),
+    verifyEmail: rateLimit({perMinute: 50, perHour: 100}),
     addApprovedOrganizationDomains: or(
       isSuperUser,
       and(isViewerBillingLeader, isOrgTier('enterprise'))
